+++ sys/src/ape/lib/fmt/fmtfdflush.c
@@ -12,6 +12,7 @@
  * OF THIS SOFTWARE OR ITS FITNESS FOR ANY PARTICULAR PURPOSE.
  */
 #include <stdarg.h>
+#include <inttypes.h>
 #include <unistd.h>
 #include "fmt.h"
 #include "fmtdef.h"
@@ -26,7 +27,7 @@ __fmtFdFlush(Fmt *f)
 	int n;
 
 	n = (char*)f->to - (char*)f->start;
-	if(n && write((int)f->farg, f->start, n) != n)
+	if(n && write((int)(uintptr_t)f->farg, f->start, n) != n)
 		return 0;
 	f->to = f->start;
 	return 1;
