+++ sys/src/cmd/eqn/lex.c
@@ -47,7 +47,8 @@ yylex(void)
 				ERROR "quoted string %.20s... too long", token FATAL;
 		}
 		token[sp] = '\0';
-		yylval = (int) &token[0];
+		/* NB: token isn't on stack, so works on amd64 */
+		yylval = (int)(uintptr)&token[0];
 		if (c == '\n')
 			ERROR "missing \" in %.20s", token WARNING;
 		return(QTEXT);
@@ -141,7 +142,7 @@ void getstr(char *s, int n)
 	}
 	unput(c);
 	*p = '\0';
-	yylval = (int) s;
+	yylval = (int)(uintptr)s;	/* hope s isn't on the stack on amd64 */
 }
 
 cstr(char *s, int quote, int maxs)
