+++ sys/src/cmd/gs/src/gsmemlok.c
@@ -177,7 +177,7 @@ gs_locked_consolidate_free(gs_memory_t * mem)
 		 );
 }
 private byte *
-gs_locked_alloc_bytes(gs_memory_t * mem, uint size, client_name_t cname)
+gs_locked_alloc_bytes(gs_memory_t * mem, uintptr_t size, client_name_t cname)
 {
     RETURN_MONITORED(
 		     byte *,
@@ -186,7 +186,7 @@ gs_locked_alloc_bytes(gs_memory_t * mem, uint size, client_name_t cname)
 		     );
 }
 private byte *
-gs_locked_alloc_bytes_immovable(gs_memory_t * mem, uint size,
+gs_locked_alloc_bytes_immovable(gs_memory_t * mem, uintptr_t size,
 				client_name_t cname)
 {
     RETURN_MONITORED(
@@ -216,7 +216,7 @@ gs_locked_alloc_struct_immovable(gs_memory_t * mem,
 		     );
 }
 private byte *
-gs_locked_alloc_byte_array(gs_memory_t * mem, uint num_elements, uint elt_size,
+gs_locked_alloc_byte_array(gs_memory_t * mem, uint num_elements, uintptr_t elt_size,
 			   client_name_t cname)
 {
     RETURN_MONITORED(
@@ -227,7 +227,7 @@ gs_locked_alloc_byte_array(gs_memory_t * mem, uint num_elements, uint elt_size,
 }
 private byte *
 gs_locked_alloc_byte_array_immovable(gs_memory_t * mem, uint num_elements,
-				     uint elt_size, client_name_t cname)
+				     uintptr_t elt_size, client_name_t cname)
 {
     RETURN_MONITORED(
 		     byte *,
@@ -292,7 +292,7 @@ gs_locked_free_object(gs_memory_t * mem, void *ptr, client_name_t cname)
 		 );
 }
 private byte *
-gs_locked_alloc_string(gs_memory_t * mem, uint nbytes, client_name_t cname)
+gs_locked_alloc_string(gs_memory_t * mem, uintptr_t nbytes, client_name_t cname)
 {
     RETURN_MONITORED(
 		     byte *,
@@ -301,7 +301,7 @@ gs_locked_alloc_string(gs_memory_t * mem, uint nbytes, client_name_t cname)
 		     );
 }
 private byte *
-gs_locked_alloc_string_immovable(gs_memory_t * mem, uint nbytes,
+gs_locked_alloc_string_immovable(gs_memory_t * mem, uintptr_t nbytes,
 				 client_name_t cname)
 {
     RETURN_MONITORED(
@@ -322,7 +322,7 @@ gs_locked_resize_string(gs_memory_t * mem, byte * data, uint old_num,
 		     );
 }
 private void
-gs_locked_free_string(gs_memory_t * mem, byte * data, uint nbytes,
+gs_locked_free_string(gs_memory_t * mem, byte * data, uintptr_t nbytes,
 		      client_name_t cname)
 {
     DO_MONITORED(
