+++ sys/src/cmd/upas/bayes/bayes.c
@@ -45,7 +45,7 @@ usage(void)
 }
 
 void*
-emalloc(int n)
+emalloc(uintptr n)
 {
 	void *v;
