+++ sys/src/cmd/usb/disk/disk.c
@@ -492,8 +492,9 @@ dread(Usbfs *fs, Fid *fid, void *data, long count, vlong offset)
 		s = buf;
 		e = buf + sizeof(buf);
 		if(lun->flags & Finqok)
-			s = seprint(s, e, "inquiry %s lun %ld: %s\n",
-				fs->dev->dir, lun - &ums->lun[0], lun->inq);
+			s = seprint(s, e, "inquiry %s lun %d: %s\n",
+				fs->dev->dir, (int)(lun - &ums->lun[0]),
+				lun->inq);
 		if(lun->blocks > 0)
 			s = seprint(s, e, "geometry %llud %ld\n",
 				lun->blocks, lun->lbsize);
