+++ sys/src/libbin/bin.c
@@ -4,8 +4,11 @@
 
 enum
 {
-	StructAlign = sizeof(union {vlong vl; double d; ulong p; void *v;
-				struct{vlong v;}vs; struct{double d;}ds; struct{ulong p;}ss; struct{void *v;}xs;})
+	StructAlign = sizeof(union {
+		vlong vl; double d; ulong p; void *v;
+		struct{vlong v;}vs; struct{double d;}ds;
+		struct{ulong p;}ss; struct{void *v;}xs;
+	})
 };
 
 enum
@@ -16,7 +19,7 @@ enum
 struct Bin
 {
 	Bin	*next;
-	ulong	total;			/* total bytes allocated in can->next */
+	uintptr	total;			/* total bytes allocated in can->next */
 	uintptr	pos;
 	uintptr	end;
 	uintptr	v;			/* last value allocated */
@@ -27,7 +30,7 @@ struct Bin
  * allocator which allows an entire set to be freed at one time
  */
 static Bin*
-mkbin(Bin *bin, ulong size)
+mkbin(Bin *bin, uintptr size)
 {
 	Bin *b;
 
@@ -45,7 +48,7 @@ mkbin(Bin *bin, ulong size)
 }
 
 void*
-binalloc(Bin **bin, ulong size, int zero)
+binalloc(Bin **bin, uintptr size, int zero)
 {
 	Bin *b;
 	uintptr p;
@@ -76,7 +79,7 @@ binalloc(Bin **bin, ulong size, int zero)
 }
 
 void*
-bingrow(Bin **bin, void *op, ulong osize, ulong size, int zero)
+bingrow(Bin **bin, void *op, uintptr osize, uintptr size, int zero)
 {
 	Bin *b;
 	void *np;
